+++ pkg/local_object_storage/engine/engine_test.go
@@ -186,7 +186,7 @@ func generateObjectWithCID(t testing.TB, cnr cid.ID) *object.Object {
 
 	obj := object.New()
 	obj.SetID(oidtest.ID())
-	owner := usertest.ID(t)
+	owner := usertest.ID()
 	obj.SetOwnerID(&owner)
 	obj.SetContainerID(cnr)
 	obj.SetVersion(&ver)