+++ pkg/local_object_storage/shard/shard_test.go
@@ -118,7 +118,7 @@ func generateObjectWithPayload(t testing.TB, cnr cid.ID, data []byte) *object.Ob
 
 	obj := object.New()
 	obj.SetID(oidtest.ID())
-	owner := usertest.ID(t)
+	owner := usertest.ID()
 	obj.SetOwnerID(&owner)
 	obj.SetContainerID(cnr)
 	obj.SetVersion(&ver)